Stop feeding exit status back to mdm-run
I have come to the conclusion that trying to pass exit status of
previous commands semi-randomly back to mdm-run is a bad idea.  In this
revision I changed mdm-master so that it always write 0 as status back
to mdm-run and removed the status field of struct slave which is no
longer needed.