Pause mdm-slave on non-zero exit status
Since mdm breaks the usual shell error-reporting mechanism, when a
parallel command exits with non-zero status, we should give the user an
opportunity to investigate the problem.  In this revision I made the
slave to wait for user acknowledgement before accepting a new command if
the previous one exited with an error.  The mdm-top monitoring program
is also changed to blink the commands with error to alert the user.