Simplify find_run semantics
Instead of finding the run with a specific PID and stage, now find_run
just tries to find the last run with the given PID.  The new behavior
should be identical to the old one in all valid use cases.  I also
removed the find_run call in start_run because it always returns rc-1.